Remove computational bottleneck from creating folds
The algorithm for creating folds is based on a matrix multiplication.
When the input data X is (items x features), this is fast. When the
input data has more dimensions, e.g. (items x channels x times), the
matrix multiplication falls back to doing a very slow loop. So, when
doing folds, we should always flatten the data first.